refactor(jsx): split runtime monolith and unify serialization paths
Extract guards, factory, template shape/compiler, and shared bindings so
jsx-runtime becomes a thin barrel. SSR, CE fallback, and hydration share
serializeRenderable and scoped subscription hand-off.
