+++ packages/jsx/src/attribute-normalize.ts
@@ -1,0 +1,119 @@
+/**
+ * Iterates normalized JSX attributes and invokes `append` for each resolved name/value
+ * pair. No intermediate record is created.
+ *
+ * @param attributes Raw JSX props object.
+ * @param append Callback invoked once per normalized attribute name/value pair.
+ */
+export function forEachNormalizedAttribute(
+	attributes: Record<string, unknown>,
+	append: (name: string, value: unknown) => void,
+): void {
+	const classValue = normalizeMergedClassValue(attributes.class, attributes.classes);
+
+	if (classValue !== undefined) {
+		append('class', classValue);
+	}
+
+	for (const name in attributes) {
+		const value = attributes[name];
+
+		if (value === undefined || name === 'key' || name === 'class' || name === 'classes') {
+			continue;
+		}
+
+		if (name === 'data' || name === 'aria') {
+			appendStructuredAttributes(name, value, append);
+			continue;
+		}
+
+		if (name === 'style') {
+			append('style', normalizeStyleValue(value));
+			continue;
+		}
+
+		append(name, value);
+	}
+}
+
+function appendStructuredAttributes(
+	prefix: 'aria' | 'data',
+	value: unknown,
+	append: (name: string, value: unknown) => void,
+): void {
+	if (!isPlainObject(value)) {
+		return;
+	}
+
+	for (const name in value) {
+		append(`${prefix}-${toKebabCase(name)}`, value[name]);
+	}
+}
+
+function normalizeMergedClassValue(classValue: unknown, classesValue: unknown): unknown {
+	const tokens: string[] = [];
+	appendClassTokens(tokens, classValue);
+	appendClassTokens(tokens, classesValue);
+	return tokens.length === 0 ? undefined : tokens.join(' ');
+}
+
+function appendClassTokens(tokens: string[], value: unknown): void {
+	if (value === undefined || value === null || value === false || value === true) {
+		return;
+	}
+
+	if (typeof value === 'string') {
+		if (value !== '') {
+			tokens.push(value);
+		}
+		return;
+	}
+
+	if (typeof value === 'number' || typeof value === 'bigint') {
+		tokens.push(String(value));
+		return;
+	}
+
+	if (Array.isArray(value)) {
+		for (const entry of value) {
+			appendClassTokens(tokens, entry);
+		}
+		return;
+	}
+
+	if (!isPlainObject(value)) {
+		return;
+	}
+
+	for (const [name, enabled] of Object.entries(value)) {
+		if (enabled) {
+			tokens.push(name);
+		}
+	}
+}
+
+function normalizeStyleValue(value: unknown): unknown {
+	if (!isPlainObject(value)) {
+		return value;
+	}
+
+	const declarations: string[] = [];
+
+	for (const [name, entry] of Object.entries(value)) {
+		if (entry === undefined || entry === null || entry === '') {
+			continue;
+		}
+
+		declarations.push(`${toKebabCase(name)}: ${String(entry)}`);
+	}
+
+	return declarations.join('; ');
+}
+
+function isPlainObject(value: unknown): value is Record<string, unknown> {
+	return typeof value === 'object' && value !== null && !Array.isArray(value);
+}
+
+function toKebabCase(value: string): string {
+	return value.replace(/[A-Z]/g, (segment) => `-${segment.toLowerCase()}`);
+}
