+++ packages/jsx/src/hydration-bindings.ts
@@ -1,43 +1,13 @@
 import type { JsxRenderable, TemplateResultLike } from './jsx-runtime.ts';
-import { isTemplateResultLike } from './jsx-runtime.ts';
+import { isIterableRenderable, isTemplateResultLike } from './renderable-guards.ts';
+import { shouldSkipHydrationSubtree } from './hydration-subtree-policy.ts';
+import { getTemplateInterpolationParts, type BindingKind } from './template-shape.ts';
+
+export type { BindingKind, TemplateInterpolationPart } from './template-shape.ts';
+export { ATTRIBUTE_BINDING_PATTERN, getBindingKind, getTemplateInterpolationParts } from './template-shape.ts';
 
 /** Attribute prefix used for emitted SSR hydration markers. */
 export const ATTRIBUTE_BINDING_PREFIX = 'data-radiant-jsx-bind-';
-/**
- * Matches a static template segment that ends with an attribute-like binding
- * placeholder such as ` class=`, ` ?hidden=`, ` @focus=`, ` !click=`, or ` .value=`.
- */
-export const ATTRIBUTE_BINDING_PATTERN = /^(.*?)(\s+)([@.?!]?)([^\s"'<>/=`]+)=$/s;
-
-/** Kinds of dynamic bindings that the JSX runtime can encode into templates. */
-export type BindingKind = 'attr' | 'bool' | 'event' | 'native-event' | 'prop';
-
-/**
- * Cached metadata describing how one interpolation slot should be interpreted
- * during SSR, hydration binding collection, or DOM template compilation.
- */
-export type TemplateInterpolationPart =
-	| {
-			/** Literal HTML prefix to write before rendering a child interpolation. */
-			string: string;
-			type: 'child';
-	  }
-	| {
-			/** Binding kind inferred from the interpolation prefix. */
-			kind: BindingKind;
-			/** Static HTML emitted before the binding marker. */
-			leading: string;
-			/** Final DOM attribute or property name. */
-			name: string;
-			/** Raw prefix from the template syntax such as `?`, `@`, or `.`. */
-			prefix: string;
-			/** Whitespace that originally separated the attribute from prior markup. */
-			whitespace: string;
-			type: 'attribute';
-	  };
-
-const TEMPLATE_INTERPOLATION_CACHE = new WeakMap<readonly string[], readonly TemplateInterpolationPart[]>();
-const TEMPLATE_INTERPOLATION_CACHE_BY_KEY = new Map<string, readonly TemplateInterpolationPart[]>();
 
 export type HydrationBinding = {
 	/** Binding category used by the hydrator. */
@@ -76,28 +46,6 @@ export function collectHydrationBindings(
 }
 
 /**
- * Resolves the logical binding kind from the runtime prefix used in JSX
- * template strings.
- *
- * @param prefix Raw prefix emitted by the JSX runtime.
- * @returns Binding category consumed by the hydrator and DOM renderer.
- */
-export function getBindingKind(prefix: string): BindingKind {
-	switch (prefix) {
-		case '!':
-			return 'event';
-		case '@':
-			return 'native-event';
-		case '.':
-			return 'prop';
-		case '?':
-			return 'bool';
-		default:
-			return 'attr';
-	}
-}
-
-/**
  * Parses a serialized hydration descriptor back into its binding metadata.
  *
  * @param value Descriptor produced by {@link serializeBindingDescriptor}.
@@ -132,67 +80,6 @@ export function serializeBindingDescriptor(kind: BindingKind, name: string): str
 	return `${kind}:${name}`;
 }
 
-/**
- * Parses the static segments for a template shape once and caches the
- * interpolation metadata by template shape.
- *
- * The cache uses the string-array identity when available and falls back to a
- * stable shape key when the runtime rebuilds the `TemplateStringsArray`. That
- * keeps server rendering and DOM compilation fast without assuming one cache
- * strategy fits every template producer.
- *
- * @param strings Static template segments for one JSX template shape.
- * @returns Parsed interpolation metadata for each dynamic slot.
- */
-export function getTemplateInterpolationParts(strings: readonly string[]): readonly TemplateInterpolationPart[] {
-	const cachedParts = TEMPLATE_INTERPOLATION_CACHE.get(strings);
-
-	if (cachedParts) {
-		return cachedParts;
-	}
-
-	const cacheKey = getTemplateCacheKey(strings);
-	const cachedPartsByKey = TEMPLATE_INTERPOLATION_CACHE_BY_KEY.get(cacheKey);
-
-	if (cachedPartsByKey) {
-		TEMPLATE_INTERPOLATION_CACHE.set(strings, cachedPartsByKey);
-		return cachedPartsByKey;
-	}
-
-	const parts: TemplateInterpolationPart[] = [];
-
-	for (let index = 0; index < strings.length - 1; index += 1) {
-		const stringPart = strings[index] ?? '';
-		const attributeBinding = ATTRIBUTE_BINDING_PATTERN.exec(stringPart);
-
-		if (!attributeBinding) {
-			parts.push({
-				string: stringPart,
-				type: 'child',
-			});
-			continue;
-		}
-
-		const [, leading = '', whitespace = ' ', prefix = '', name = ''] = attributeBinding;
-		parts.push({
-			kind: getBindingKind(prefix),
-			leading,
-			name,
-			prefix,
-			whitespace,
-			type: 'attribute',
-		});
-	}
-
-	TEMPLATE_INTERPOLATION_CACHE.set(strings, parts);
-	TEMPLATE_INTERPOLATION_CACHE_BY_KEY.set(cacheKey, parts);
-	return parts;
-}
-
-function getTemplateCacheKey(strings: readonly string[]): string {
-	return strings.map((part) => `${part.length}:${part}`).join('|');
-}
-
 function collectValueBindings(
 	value: JsxRenderable,
 	bindings: Map<number, HydrationBinding>,
@@ -208,7 +95,7 @@ function collectValueBindings(
 		return;
 	}
 
-	if (isIterable(value)) {
+	if (isIterableRenderable(value)) {
 		for (const child of value) {
 			collectValueBindings(child as JsxRenderable, bindings, state, options);
 		}
@@ -249,10 +136,7 @@ function collectTemplateBindings(
 function isCustomElementTemplateRoot(template: TemplateResultLike): boolean {
 	const openingSegment = template.strings[0]?.trimStart() ?? '';
 	const tagMatch = /^<([a-z][\w.-]*)\b/i.exec(openingSegment);
+	const tagName = tagMatch?.[1];
 
-	return tagMatch?.[1]?.includes('-') === true;
-}
-
-function isIterable(value: unknown): value is Iterable<unknown> {
-	return typeof value !== 'string' && typeof value === 'object' && value !== null && Symbol.iterator in value;
+	return tagName !== undefined && shouldSkipHydrationSubtree({ localName: tagName } as Element);
 }
